add create-skill-form component
