+++ src/presentation/pages/create-skill.module.css
@@ -4,11 +4,3 @@
   justify-content: center;
   align-items: center;
 }
-
-.CreateSkill__input {
-  width: 15rem;
-}
-.CreateSkill__btn {
-  width: 8rem;
-  margin: 1rem;
-}
