+++ src/presentation/pages/create-skill.tsx
@@ -1,42 +1,16 @@
-import { FC, FormEvent } from "react";
+import { FC } from "react";
 import { errorActions } from "../../services/redux/errorSlice";
-import CreateButton from "../components/atoms/create-button";
-import Input from "../components/atoms/input";
-import { useAppDispatch, useAppSelector } from "../hooks/reduxHooks";
-import { useSkillMutations } from "../hooks/useSkillMutations";
-import useForm from "../hooks/useForm";
+import { useAppDispatch } from "../hooks/reduxHooks";
 import useSkillsQuery from "../hooks/useSkillsQuery";
 import classes from "./create-skill.module.css";
 import SkillCard from "../components/organisms/SkillCard";
-
-const formInitialValues = {
-  title: "",
-};
+import CreateSkillForm from "../components/organisms/CreateSkillForm";
 
 const CreateSkill: FC = () => {
   const dispatch = useAppDispatch();
-  const { token } = useAppSelector((state) => state.auth);
   const skillsQuery = useSkillsQuery();
-  const { createSkillMutation } = useSkillMutations();
+
   const { data: skills } = skillsQuery;
-
-  const { values, valueChangeHandler, resetValues } =
-    useForm(formInitialValues);
-  const { title } = values;
-
-  const formSubmitHandler = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    if (!token) {
-      return dispatch(
-        errorActions.setError(`Failed to create skill: No token found`)
-      );
-    }
-
-    createSkillMutation.mutate({ newSkill: { title }, token });
-
-    resetValues();
-  };
 
   if (skillsQuery.isLoading) {
     return <div>Loading...</div>;
@@ -53,34 +27,9 @@ const CreateSkill: FC = () => {
     );
   }
 
-  if (createSkillMutation.isError) {
-    console.log("skillError", createSkillMutation.error);
-
-    dispatch(
-      errorActions.setError(
-        `Failed to create skill: ${createSkillMutation.error as Error}`
-      )
-    );
-
-    createSkillMutation.reset();
-  }
-
   return (
     <div className={classes[componentName]}>
-      <form onSubmit={formSubmitHandler}>
-        {/* FIXME: replace with mui-v5 component */}
-        <Input
-          name="title"
-          placeholder="Title"
-          className={classes[`${componentName}__input`]}
-          value={title}
-          onChange={valueChangeHandler}
-        />
-        <CreateButton
-          title="Add Skill"
-          className={classes[`${componentName}__btn`]}
-        />
-      </form>
+      <CreateSkillForm />
       <div className="grid grid-cols-4 gap-4">
         {skills?.map((skill) => (
           <SkillCard key={skill._id} skill={skill} />
